flutter_car_app/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
werkzeug==2.3.7
numpy==1.24.3 